fixes config sccp_group_t issue
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@2828 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/chan_sccp.h
@@ -193,7 +193,7 @@ extern "C" {
 	#define SCCP_MAX_MUSICCLASS 80
 #endif
 
-	typedef unsigned int sccp_group_t;
+	typedef unsigned long long sccp_group_t;
 	typedef struct channel sccp_channel_t;					/*!< SCCP Channel Structure */
 	typedef struct sccp_session sccp_session_t;				/*!< SCCP Session Structure */
 	typedef struct sccp_line sccp_line_t;					/*!< SCCP Line Structure */
